refactor(transcode): relocate TranscodeService to core and flatten API package
