+++ packages/api/src/file.ts
@@ -17,7 +17,7 @@ import {
   uploadFileRequestSchema,
   getDownloadLinksRequestSchema,
 } from '@shumai/dtos'
-import { transcodeService } from '@shumai/transcode/src/transcode'
+import { transcodeService } from '@shumai/core/src/transcode/transcode'
 import fs from 'fs'
 import { Hono } from 'hono'
 import os from 'os'
