+++ packages/api/src/index.ts
@@ -1,25 +1,25 @@
 import type { Prisma } from '@shumai/db'
 import { Hono } from 'hono'
-import agentRoute from './api/agent'
-import attachmentRoute from './api/attachment'
-import authnRoute from './api/authn'
-import collectionRoute from './api/collection'
-import fileRoute from './api/file'
-import folderRoute from './api/folder'
-import inviteRoute from './api/invite'
-import metadataRoute from './api/metadata'
-import { authMiddleware } from './api/middleware/auth'
-import notificationRoute from './api/notification'
-import projectRoute from './api/project'
-import providerRoute from './api/provider'
-import publicInviteRoute from './api/public-invite'
-import publicShareRoute from './api/public-share'
-import s3Route from './api/s3'
-import shareRoute from './api/share'
-import skillRoute from './api/skill'
-import teamRoute from './api/team'
-import uploadRoute from './api/upload'
-import versionStackRoute from './api/versionStack'
+import agentRoute from './agent'
+import attachmentRoute from './attachment'
+import authnRoute from './authn'
+import collectionRoute from './collection'
+import fileRoute from './file'
+import folderRoute from './folder'
+import inviteRoute from './invite'
+import metadataRoute from './metadata'
+import { authMiddleware } from './middleware/auth'
+import notificationRoute from './notification'
+import projectRoute from './project'
+import providerRoute from './provider'
+import publicInviteRoute from './public-invite'
+import publicShareRoute from './public-share'
+import s3Route from './s3'
+import shareRoute from './share'
+import skillRoute from './skill'
+import teamRoute from './team'
+import uploadRoute from './upload'
+import versionStackRoute from './versionStack'
 
 type User = Prisma.UserGetPayload<Record<string, never>>
