+++ packages/api/src/middleware/auth.ts
