+++ packages/core/package.json
@@ -18,6 +18,7 @@
     "hono": "4.12.23",
     "jittered-fractional-indexing": "1.0.1",
     "pino": "10.3.1",
+    "sharp": "0.34.5",
     "tweetnacl": "1.0.3",
     "ulid": "3.0.2",
     "yaml": "2.9.0",
