+++ packages/core/src/index.ts
@@ -13,6 +13,7 @@ export * from './search/search'
 export * from './share/share'
 export * from './skill/skill'
 export * from './team/team'
+export * from './transcode/transcode'
 export * from './upload/upload'
 export * from './user-metadata/user-metadata'
 export * from './user/avatar'
