+++ packages/core/src/transcode/dataFormatNames.ts
@@ -1,0 +1,115 @@
+export const dataFormatNames: Record<string, string> = {
+  // additional codecs
+  apch: 'Apple ProRes 422 High Quality',
+  apcn: 'Apple ProRes 422 Standard Definition',
+  apcs: 'Apple ProRes 422 LT',
+  apco: 'Apple ProRes 422 Proxy',
+  ap4h: 'Apple ProRes 4444',
+  jpeg: 'JPEG Image',
+
+  // codecs from https://mp4ra.org/
+  '3gvo': '3GPP Video Orientation',
+  a3d1: 'Multiview Video Coding',
+  a3d2: 'Multiview Video Coding',
+  a3d3: 'Multiview Video Coding',
+  a3d4: 'Multiview Video Coding',
+  a3ds: 'Auro-Cx 3D audio',
+  'ac-3': 'AC-3 audio',
+  'ac-4': 'AC-4 audio',
+  alac: 'Apple lossless audio codec',
+  alaw: 'a-Law',
+  av01: 'AV1 video',
+  avc1: 'Advanced Video Coding',
+  avc2: 'Advanced Video Coding',
+  avc3: 'Advanced Video Coding',
+  avc4: 'Advanced Video Coding',
+  avcp: 'Advanced Video Coding Parameters',
+  dra1: 'DRA Audio',
+  drac: 'Dirac Video Coder',
+  'dts-': 'Dependent base layer for DTS layered audio',
+  'dts+': 'Enhancement layer for DTS layered audio',
+  dtsc: 'DTS Coherent Acoustics audio',
+  dtse: 'DTS Express low bit rate audio, also known as DTS LBR',
+  dtsh: 'DTS-HD High Resolution Audio',
+  dtsl: 'DTS-HD Master Audio',
+  dtsx: 'DTS:X',
+  dvav: 'AVC-based “Dolby Vision”',
+  dvhe: 'HEVC-based “Dolby Vision”',
+  'ec-3': 'Enhanced AC-3 audio',
+  enca: 'Encrypted/Protected audio',
+  encf: 'Encrypted/Protected font',
+  encm: 'Encrypted/Protected metadata stream',
+  encs: 'Encrypted Systems stream',
+  enct: 'Encrypted Text',
+  encv: 'Encrypted/protected video',
+  fdp: 'File delivery hints',
+  fLaC: 'Fres Lossless Audio Codec',
+  g719: 'ITU-T Recommendation G.719 (2008)',
+  g726: 'ITU-T Recommendation G.726 (1990)',
+  hev1: 'High Efficiency Video Coding',
+  hvc1: 'High Efficiency Video Coding',
+  hvt1: 'High Efficiency Video Coding',
+  ixse: 'DVB Track Level Index Track',
+  lhe1: 'Layered High Efficiency Video Coding',
+  lht1: 'Layered High Efficiency Video Coding',
+  lhv1: 'Layered High Efficiency Video Coding',
+  m2ts: 'MPEG-2 transport stream for DMB',
+  m4ae: 'MPEG-4 Audio Enhancement MP4v1/2',
+  mett: 'Text timed metadata that is not XML',
+  metx: 'XML timed metadata',
+  mha1: 'MPEG-H Audio (single stream, uncapsulated)',
+  mha2: 'MPEG-H Audio (multi-stream, unencapsulated)',
+  mhm1: 'MPEG-H Audio (single stream, MHAS encapsulated)',
+  mhm2: 'MPEG-H Audio (multi-stream, MHAS encapsulated)',
+  mjp2: 'Motion JPEG 2000',
+  mlix: 'DVB Movie level index track',
+  mlpa: 'MLP Audio',
+  mp4a: 'MPEG-4 Audio',
+  mp4s: 'MPEG-4 Systems',
+  mp4v: 'MPEG-4 Visual',
+  mvc1: 'Multiview coding',
+  mvc2: 'Multiview coding',
+  mvc3: 'Multiview coding',
+  mvc4: 'Multiview coding',
+  mvd1: 'Multiview coding',
+  mvd2: 'Multiview coding',
+  mvd3: 'Multiview coding',
+  mvd4: 'Multiview coding',
+  oksd: 'OMA Keys',
+  // eslint-disable-next-line @typescript-eslint/naming-convention
+  Opus: 'Opus audio coding',
+  pm2t: 'Protected MPEG-2 Transport',
+  prtp: 'Protected RTP Reception',
+  raw: 'Uncompressed audio',
+  resv: 'Restricted Video',
+  rm2t: 'MPEG-2 Transport Reception',
+  rrtp: 'RTP reception',
+  rsrp: 'SRTP Reception',
+  rtmd: 'Real Time Metadata Sample Entry(XAVC Format)',
+  rtp: 'RTP Hints',
+  s263: 'ITU H.263 video (3GPP format)',
+  samr: 'Narrowband AMR voice',
+  sawb: 'Wideband AMR voice',
+  sawp: 'Extended AMR-WB (AMR-WB+)',
+  sevc: 'EVRC Voice',
+  sm2t: 'MPEG-2 Transport Server',
+  sqcp: '13K Voice',
+  srtp: 'SRTP Hints',
+  ssmv: 'SMV Voice',
+  STGS: 'Subtitle Sample Entry (HMMP)',
+  stpp: 'Subtitles (Timed Text)',
+  svc1: 'Scalable Video Coding',
+  svc2: 'Scalable Video Coding',
+  svcM: 'SVC Metadata',
+  tc64: '64 bit timecode samples',
+  tmcd: '32 bit timecode samples',
+  twos: 'Uncompressed 16-bit audio',
+  tx3g: 'Timed Text stream',
+  ulaw: 'Samples have been compressed using uLaw 2:1.',
+  unid: 'Dynamic Range Control (DRC) data',
+  urim: 'Binary timed metadata identified by URI',
+  'vc-1': 'SMPTE VC-1',
+  vp08: 'VP8 video',
+  vp09: 'VP9 video',
+  wvtt: 'WebVTT',
+}
