+++ packages/transcode/package.json
@@ -13,7 +13,6 @@
     "@temporalio/activity": "1.17.2",
     "@temporalio/worker": "1.17.2",
     "@temporalio/workflow": "1.17.2",
-    "sharp": "0.34.5",
     "ulid": "3.0.2"
   }
 }
